# backend/assets/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "# Nexus Tasks REST API\n\nNexus Tasks is a collaborative project and task management platform.\nThis document describes the full REST API surface of the backend service.\n\n## Authentication\n\nMost endpoints require an authenticated session established via the `/api/v2/auth/login`\nor `/api/v2/auth/register` endpoints. Authentication state is maintained via an\nHTTP-only session cookie (`nexus_user`). Alternatively, machine clients may\nauthenticate using an API Key passed in the `X-API-Key` request header.\n\n## Base URL\n\nAll versioned resources are prefixed with `/api/v2`.",
    "title": "provides the Swagger/OpenAPI specification metadata for the Nexus Tasks API.",
    "version": "2.0.0"
  },
  "host": "localhost:8000",
  "basePath": "/api/v2",
  "paths": {
...
    "/workspaces": {
      "get": {
        "description": "Returns all workspaces the authenticated user is a member of.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "workspaces"
        ],
        "summary": "List workspaces",
        "operationId": "listWorkspaces",
        "responses": {
          "200": {
            "description": "List of workspaces.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ResWorkspace"
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "description": "Creates a new workspace owned by the authenticated user.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "workspaces"
        ],
        "summary": "Create a workspace",
        "operationId": "createWorkspace",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReqCreateWorkspace"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Workspace created.",
            "schema": {
              "$ref": "#/definitions/ResWorkspace"
            }
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/workspaces/{workspaceId}/members": {
      "get": {
        "description": "Returns all users who are members of the given workspace.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "workspaces"
        ],
        "summary": "List workspace members",
        "operationId": "listWorkspaceMembers",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the workspace.",
            "name": "workspaceId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "List of workspace members.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ResWorkspaceMember"
              }
            }
          },
          "400": {
            "description": "Invalid workspace ID."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "description": "Sends an invitation (or directly adds) a user to the workspace by email address.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "workspaces"
        ],
        "summary": "Invite a user to a workspace",
        "operationId": "inviteWorkspaceMember",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the workspace.",
            "name": "workspaceId",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReqInviteWorkspaceMember"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member invited successfully."
          },
          "400": {
            "description": "Validation error or invalid workspace ID."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "description": "Removes the specified user from the workspace. Only workspace admins may perform this action.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "workspaces"
        ],
        "summary": "Remove a member from a workspace",
        "operationId": "removeWorkspaceMember",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the workspace.",
            "name": "workspaceId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the user to remove.",
            "name": "uid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Member removed successfully."
          },
          "400": {
            "description": "Invalid workspace or user ID."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "ReqCreateWorkspace": {
      "type": "object",
      "title": "ReqCreateWorkspace defines the request payload for creating a workspace.",
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "description": "Workspace display name.",
          "type": "string",
          "x-go-name": "Name",
          "example": "Acme Corp"
        }
      },
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/pkg/structs"
    },
...
    "ReqInviteWorkspaceMember": {
      "type": "object",
      "title": "ReqInviteWorkspaceMember defines the request payload for inviting a member to a workspace.",
      "required": [
        "email"
      ],
      "properties": {
        "email": {
          "description": "Email of the user to invite.",
          "type": "string",
          "x-go-name": "Email",
          "example": "colleague@example.com"
        }
      },
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/pkg/structs"
    },
...
    "ResUser": {
      "type": "object",
      "title": "ResUser defines the response payload for a user.",
      "properties": {
        "email": {
          "description": "User's email address.",
          "type": "string",
          "x-go-name": "Email",
          "example": "user@example.com"
        },
        "full_name": {
          "description": "User's display name.",
          "type": "string",
          "x-go-name": "FullName",
          "example": "Jane Doe"
        },
        "id": {
          "description": "Unique user identifier.",
          "type": "string",
          "format": "uuid",
          "x-go-name": "ID",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        },
        "is_active": {
          "description": "Whether the user account is active.",
          "type": "boolean",
          "x-go-name": "IsActive",
          "example": true
        }
      },
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/pkg/structs"
    },
    "ResWorkspace": {
      "type": "object",
      "title": "ResWorkspace defines the response payload for a workspace.",
      "properties": {
        "id": {
          "description": "Unique workspace identifier.",
          "type": "string",
          "format": "uuid",
          "x-go-name": "ID",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        },
        "name": {
          "description": "Workspace display name.",
          "type": "string",
          "x-go-name": "Name",
          "example": "Acme Corp"
        },
        "owner_id": {
          "description": "UUID of the workspace owner.",
          "type": "string",
          "format": "uuid",
          "x-go-name": "OwnerID",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        },
        "type": {
          "$ref": "#/definitions/WorkspaceType"
        }
      },
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/pkg/structs"
    },
    "ResWorkspaceMember": {
      "type": "object",
      "title": "ResWorkspaceMember defines the response payload for a workspace member.",
      "properties": {
        "role": {
          "$ref": "#/definitions/WorkspaceRole"
        },
        "user": {
          "$ref": "#/definitions/ResUser"
        },
        "user_id": {
          "description": "UUID of the member user.",
          "type": "string",
          "format": "uuid",
          "x-go-name": "UserID",
          "example": "550e8400-e29b-41d4-a716-446655440001"
        },
        "workspace_id": {
          "description": "UUID of the workspace.",
          "type": "string",
          "format": "uuid",
          "x-go-name": "WorkspaceID",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        }
      },
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/pkg/structs"
    },
...
    "WorkspaceRole": {
      "description": "WorkspaceRole represents the role of a user in a workspace",
      "type": "string",
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/models"
    },
    "WorkspaceType": {
      "description": "WorkspaceType represents the type of a workspace",
      "type": "string",
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/models"
    },
...
  },
...
  "securityDefinitions": {
    "apiKeyAuth": {
      "description": "A long-lived API key generated via the `/api/v2/auth/api-keys` endpoint. Intended for machine-to-machine access.",
      "type": "apiKey",
      "name": "X-API-Key",
      "in": "header"
    },
    "cookieAuth": {
      "description": "HTTP-only session cookie set by the login/register endpoints. The cookie name is `nexus_user` and contains a signed JWT.\n",
      "type": "apiKey",
      "name": "Cookie",
      "in": "header"
    }
  },
  "security": [
    {
      "cookieAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ]
}